+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Head
+
+- Fix handling of `^` (exponentiation) and `%` (remainder) expression operators.
+
 ## 0.2.0
 
 - **[BREAKING]** Export 2 functions: `expressionToFormula` and `formulaToExpression`, so the transformation can go both ways.